added reference to parent in order to manipulate inset offsets from widgets
i know it's not "the react way," but I think it will be a better way to deal with custom needs of widgets on an individual basis. As it is, we are imperatively manipulating the underlying layout as needed. 

It will also basically allow for only having custom widgets in application code, and not having to maintain a fork that is very different from the main library.